move components to Business namespace
