+++ app/components/businesses/legacy_business_subscription_component.rb
@@ -1,0 +1,13 @@
+module Businesses
+  class LegacyBusinessSubscriptionComponent < ApplicationComponent
+    attr_reader :user
+
+    def initialize(user)
+      @user = user
+    end
+
+    def render?
+      Businesses::Permission.new(user&.subscriptions).legacy_subscription?
+    end
+  end
+end
