--- app/components/expired_subscription_component.rb
@@ -1,16 +1,0 @@
-class ExpiredSubscriptionComponent < ApplicationComponent
-  attr_reader :user, :business
-
-  def initialize(user, business:)
-    @user = user
-    @business = business
-  end
-
-  def render_content?
-    if user.business == business
-      Businesses::Permission.new(user.subscriptions).active_subscription?
-    else
-      true
-    end
-  end
-end
