--- app/components/legacy_business_subscription_component.rb
@@ -1,11 +1,0 @@
-class LegacyBusinessSubscriptionComponent < ApplicationComponent
-  attr_reader :user
-
-  def initialize(user)
-    @user = user
-  end
-
-  def render?
-    Businesses::Permission.new(user&.subscriptions).legacy_subscription?
-  end
-end
