--- app/components/upgrade_account_component.rb
@@ -1,11 +1,0 @@
-class UpgradeAccountComponent < ApplicationComponent
-  attr_reader :user
-
-  def initialize(user)
-    @user = user
-  end
-
-  def render_content?
-    Businesses::Permission.new(user.subscriptions).active_subscription?
-  end
-end
