+++ app/views/pricing/_v2.html.erb
@@ -8,7 +8,7 @@
     </div>
   </div>
 
-  <%= render LegacyBusinessSubscriptionComponent.new(current_user) do %>
+  <%= render Businesses::LegacyBusinessSubscriptionComponent.new(current_user) do %>
     <%= render "banner" %>
   <% end %>
