+++ config/locales/en.yml
@@ -110,6 +110,9 @@ en:
   businesses:
     create:
       created: Your business was added!
+    expired_subscription_component:
+      cta: Upgrade your account to message developers
+      title: Your subscription expired.
     form:
       avatar: Your company logo
       business:
@@ -141,6 +144,9 @@ en:
       title: Business subscription
     update:
       updated: Your business was updated!
+    upgrade_account_component:
+      cta: Upgrade your account
+      title: Requires an active business subscription.
   cold_messages:
     new:
       title: Start a conversation
@@ -262,9 +268,6 @@ en:
       content_type_invalid: image format not supported
       max_file_size_invalid: image over %{size}
     unauthorized: You don't have access to that.
-  expired_subscription_component:
-    cta: Upgrade your account to message developers
-    title: Your subscription expired.
   hiring_fee_component:
     cta: Pay hiring fee
     title: Have you hired %{developer}?
@@ -532,9 +535,6 @@ en:
     million: M
     thousand: K
     unit: ''
-  upgrade_account_component:
-    cta: Upgrade your account
-    title: Requires an active business subscription.
   upgrade_required_component:
     body:
       expired: Your subscription has expired. Reactivate your account to continue messaging developers.
