+++ test/components/businesses/open_graph_tags_component_test.rb
